requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
motor
pymongo
langchain
langchain-community
langchain-huggingface
langgraph
faiss-cpu
sentence-transformers
pypdf
pandas
google-generativeai
ollama
biopython
requests
reportlab
pytest
pytest-asyncio